Time value monitoring 7/requirements.txt
streamlit
pandas
fubon-neo

plotly
firebase-admin
python-dotenv

yfinance
streamlit-autorefresh
